Fix 'make dist' when pom.xml is a link
There is a bug in Maven assembly plugin that causes SOME paths
to be resolved incorrectly if pom.xml is a symlink.

Fortunately files specified in a slightly different way are
unaffected.

Note that this commit changes a release (i.e. non-snapshot)
version. We generally want to avoid that, but, in this case,
we know there's been no public release because of the bug.

Change-Id: I5cc9433e61034d337b5cfb9605cf034d9d626e84
Reviewed-on: https://gerrit.franz.com:9080/7450
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>