Move from Sesame to rdf4j
The changes are mostly package and class renames. In some
cases the replacement has a slightly different interface,
but the changes should not be visible to the users.

One externally visible change, mentioned in the migration
guide, is that RDF4J now includes the graph component
when comparing statements. A duplicate filtering test
had to be adjusted to take this into account.

Renamed META-INF entries used to register our classes
with RDF4J registries.

Added a new dependency - rdf4j-model - it contains some
code that was apparently factored out from other RDF4J
artifacts.

Adjusted the tutorial to refer to RDF4J instead of Sesame,
but the client is still called "Java Sesame API" for now.

Many methods returning RDFFromats now return an Option
object. These methods used to simply return null if
the desired format could not be found. Thus I've
appended .orElse(null) to such calls, except where
I could simplify the code by directly providing a
default value for orElse().

Replaced try-finally blocks with try-with-resources
in parts of code that had to be modified anyway.

While the code compiles and passes the tests after
this change, it should be noted that it generates
a lot of deprecation warnings that will have to be
handled at some point.

Change-Id: I0e0fbfed2197e71e1079b733dd76ff78cdf4fe00
Reviewed-on: https://gerrit.franz.com:9080/7707
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>