Optionally buffer AGRepositoryConnection.add() for efficiency
Add option to speed up repeated calls to AGRepositoryConnection.add(<statement>)
by buffering the statements, then process them in batches.

This significantly speeds up benchmarking loops that take the form:

  try (RepositoryConnection repCon = repo.getConnection()) {
    repCon.begin();
    for (int i = 0; i < nrOfStatements; i++) {
      repCon.add(vf.createStatement(subject, predicate, object, context));
    }
    repCon.commit();
  }

because the code becomes functionally equivalent to collecting the statements in a list, then adding that list:

  try (RepositoryConnection repCon = repo.getConnection()) {
    repCon.begin();
    for (int i = 0; i < nrOfStatements; i++) {
      statementList.add(vf.createStatement(subject, predicate, object, context));
    }
    repCon.add(statementList);
    repCon.commit();
  }

The trigger to handle buffered statements is either
  1) when the buffer limit is reached, or
  2) when on AGRepositoryConnection a method is called that does HTTP repository interaction,
     and that method is not repConn.add(<statement>)

The buffer will never be used in autocommit mode.

A drawback is that errors regarding the adding are not signalled until the buffer is handled.

The buffering is enabled by setting property:

  com.franz.agraph.repository.AGRepositoryConnection.useAddStatementBuffer=true

or by calling:

  AGRepositoryConnection.setAddStatementBufferEnabled(true);

The maximum buffer size is by default 10000 statement, can be set by setting property:

  com.franz.agraph.repository.AGRepositoryConnection.addStatementBufferMaxSize=10000

or by calling:

  AGRepositoryConnection.setAddStatementBufferMaxSize(n)

Rename conn.getHttpRepoClient() to conn.prepareHttpRepoClient() for clarity:
its side effect is sending over the buffered statements, if any.

Change-Id: I0a9b9402ba9e22d51c0664fa9cd4827c2e87f56d
Reviewed-on: https://gerrit.franz.com:9080/8830
Reviewed-by: Willem Broekema <metawilm@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>