Fixed a concurrency issue introduced in #8918
The problem was that after this change pooled connections used a
single-threaded HTTP client. That did not work because of the pinger.
It needs to be able to access the HTTP client concurrently with
other connection activity.

So now each connection will again have its own, multi-threaded client.

I tried creating a single client for the whole pool, but that was
problematic because it's not easy to detect when the client should
be closed.

Change-Id: I9ba720af7dbf0092ced27781e34380e856bf748f
Reviewed-on: https://gerrit.franz.com:9080/9005
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>