rfe15689: add 'warmup' to the Java client.
See RELEASE-HISTORY.md for details.

Also added the ability to retrieve the log file from the server.
It is used by tests to check if the warmup really happened.
On 6.2.3 (where warmups are not logged) the tests just check
if initiating a warmup does not cause an error.

Also made it possible to actually create a connection pool
without the 'session' parameter.

Change-Id: I3521b4d8d388402e6e4b311485d74afd5d503739
Reviewed-on: https://gerrit.franz.com:9080/10990
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>