bug25360: HTTP connection leak
Reworked the executeMethod logic to ensure that
releaseConnection() is always called on the 'method'
object, unless we can guarantee that a handler will
take responsibility for that.

Previously the connection was not released if the handler
declared that it wants to take ownership of the connection,
but never received it due to an error.

The bug resulted in warnings about HttpConnectionManager
being null shown during tests. That was caused by the client
object being released while some of its connections are still
"in use" (or actually aren't, but haven't been released
properly)

Change-Id: I1c1af5be85e49e1fd1d5d6514b6e8b875ad58793
Reviewed-on: https://gerrit.franz.com:9080/11124
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>