Use a different approach to working around the /dev/shm problem
Travis CI build started failing on Aug 7, 2019.  It looks like prior
builds were running on Ubuntu 14.04 and now they're running on Ubuntu
16.04.

The prior approach to the /dev/shm workaround was to delete /dev/shm
and recreate it as a symlink to /run/shm.  This now results in an
error:

rm: cannot remove '/dev/shm': Device or resource busy

Now we rename /dev/shm out of the way instead of trying to delete it.

Change-Id: Icfbc0b5b678cfe66943cb504e5d1723e6c45abd4
Reviewed-on: https://gerrit.franz.com:9080/13823