bug25881: update .travis.yml to enable Github CI builds
This commit removes a /dev/shm hack that used to work on older Travis
infrastructure and is not needed anymore. It also updates to the
matrix to include 6.5.0 and 6.6.0 AG releases and normalizes the YAML
formatting.

We also fix JDK version to the one that is supported by Travis and
supports -html4 flag, bump the maven-javadoc-plugin version to be able
to specify the -html4 option and use Maven's --batch-mode option to
limit the log size.

Tests added for:       no
make test-suite run?   no
Documentation changes? no

Change-Id: I71ef6e7d228d387ebafc608c854da8b68d7ba1a2
Reviewed-on: https://gerrit.franz.com:9080/14116
Reviewed-by: Yuri Zhykin <yuriz@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>