Replace deprecated HTML tags with HTML5-compliant ones
Since releases are build with JDK 1.8.0 and Github CI is run with JDK
10, the only way to use HTML4 in both cases is to use different Maven
profiles, but this will result in duplication of large chunk of
configuration, so it makes sense to just convert the conflicting tags
to HTML5-compliant ones which appear to be working fine in both cases.

Tests added for:       no
make test-suite run?   no
Documentation changes? yes, HTML tags only

Change-Id: I0c3c22f7a08e0e402f6aca9708f2ed00a42bf7b6
Reviewed-on: https://gerrit.franz.com:9080/14124
Reviewed-by: Yuri Zhykin <yuriz@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>