Track transaction state in AGRepositoryConnection
Introduce AGTransactionState enum to track the state of a transaction
based on calls of begin(), commit(), and rollback() methods.

This allows to correctly implement the behavior of
RepositoryConnection.isActive():

  try (RepositoryConnection conn = repo.getConnection()) {

    // Before begin() is called, conn.isActive() must return false,
    // but it returned true because AGHttpRepoClient.autoCommit
    // is true by default.

    conn.begin();
    ...
    // In between begin() and commit() / rollback(),
    // isActive() returned true as expected.
    ...
    con.commit();

    // After commit() or rollback(), isActive() must return false,
    // but it returned true, because autoCommit was set to false
    // by a begin() call.
  }

Autocommit-related interface has been marked as deprecated in RDF4J
for a while. To my understanding, this means that the transactional
interface (begin/commit) is recommended.

However it seems risky to change the default autoCommit behavior of
the AG client. To avoid it while ensuring the correctness of
isActive(), two distinct transaction states are introduced:
AUTOCOMMIT_DEFAULT and AUTOCOMMIT_ENABLED.

If setAutoCommit() or begin() are never explicitly called, then
isActive() will always return false, because transaction state remains
AUTOCOMMIT_DEFAULT. The assumption is that there is no reason for a
user to call isActive() if they are relying on implicit autocommit.

On the other hand, if the transactional interface is used, or if
setAutoCommit(true) is called explicitly, then the transaction state
is tracked appropriately.

Change-Id: I6a4223390418ff49d3f0b409cd97942df62accb8
Reviewed-on: https://gerrit.franz.com:9080/c/agraph-java-client/+/25974
Reviewed-by: Kevin Layer <layer@franz.com>
Reviewed-by: Ihor Solodrai <solodrai@franz.com>
Ptest: Hudson <hudson@franz.com>
Tested-by: Hudson <hudson@franz.com>